refactor: transform the single package to a workspace
Move sources to the package "client".
Create the "server" package.
Create the "common" library package and move code that will be used in
the server to it.
Add clap and prepare the client for multiplayer.
Refactor sprite module to prepare the client multiplayer (and move it to
the "common" package).